requirements.txt
pynput>=1.7.6
pyperclip>=1.8.2
pywin32>=306
pystray>=0.19.5
Pillow>=9.0.0
